agents2/requirements.txt
fastapi==0.115.11
langfuse==2.59.7
openai-agents==0.0.4
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1